requirements.txt
Flask
pymongo
scikit-learn
pandas
numpy
flask-cors
pyserial
requests